+++ src/components/TodoInput.vue
@@ -1,13 +1,69 @@
 <template>
-  <div>input</div>
+  <div class="inputBox shadow container">
+    <input type="text" ref="todo" v-model="newTodoItem" placeholder="할 일을 입력하세요."
+            v-on:keyup.enter="addTodo">
+    <span class="addContainer" vo-on:click="addTodo">
+        <i class="addBtn fas fa-plus" aria-hidden="true"></i>
+    </span>
+  </div>
 </template>
 
 <script>
     export default {
-
+        data(){
+            return {
+                newTodoItem: '',
+            }
+        },
+        methods:{
+            addTodo(){
+                if (this.newTodoItem !== ""){
+                    var value = this.newTodoItem && this.newTodoItem.trim();
+                    localStorage.setItem(value, value);
+                    this.clearInput();
+                }else{
+                    alert("할 일을 적어주세요!")
+                    this.$refs.todo.focus();
+                }
+            },
+            clearInput(){
+                this.newTodoItem = '';
+            }
+        }
     }
 </script>
 
 <style>
+  input:focus{
+    outline: none;
+  }
 
+  .inputBox{
+    text-align: center;
+    background: white;
+    height: 50px;
+    width: 400px;
+    line-height: 50px;
+    border-radius: 5px;
+    margin: 0 auto;
+  }
+
+  .inputBox input{
+    border-style: none;
+    font-size: 0.9rem;
+    width: 300px;
+    margin-left: 30px;
+  }
+
+  .addContainer{
+    float: right;
+    background: linear-gradient(to right, #6478FB, #8763FB);
+    display: block;
+    width: 3rem;
+    border-radius: 0 5px 5px 0;
+  }
+  .addBtn{
+    color: white;
+    vertical-align: middle;
+  }
 </style>